% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{load_alabaster_from_proj}
\alias{load_alabaster_from_proj}
\title{Load SingleCellExperiment Files from a single project path}
\usage{
load_alabaster_from_proj(proj_dir, ...)
}
\arguments{
\item{proj_dir}{project directory}

\item{...}{extra args passed to load_alabaster_path}
}
\value{
a SingleCellExperiment object
}
\description{
Load SingleCellExperiment Files from a single project path
}
