% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{subset_by_colData}
\alias{subset_by_colData}
\title{Subset by new colData}
\usage{
subset_by_colData(colData_path, object)
}
\arguments{
\item{colData_path}{Path to new colData}

\item{object}{A object}
}
\value{
a SingleCellExperiment object
}
\description{
Subset the object using new colData
}
