% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unite_metadata}
\alias{unite_metadata}
\title{Unite metadata}
\usage{
unite_metadata(object, group_bys)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{group_bys}{A feature or variable to combine}
}
\value{
a SingleCellExperiment object with Idents formed from concatenation of group_bys
}
\description{
Unite metadata
}
\examples{


data(small_example_dataset)
unite_metadata(small_example_dataset, "Mutation_Status")

}
