% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transcripts_from_sce}
\alias{get_transcripts_from_sce}
\title{Get Transcripts in object}
\usage{
get_transcripts_from_sce(object, gene)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{gene}{Gene of interest}
}
\value{
transcripts constituting a
gene of interest in a SingleCellExperiment object
}
\description{
Get transcript ids in objects for one or more gene of interest
}
